+++ catboost/private/libs/algo_helpers/approx_calcer_helpers.cpp
@@ -1,8 +1,8 @@
 #include "approx_calcer_helpers.h"
 
 double CalcSampleQuantile(
-    TConstArrayRef<double> sample,
-    TConstArrayRef<double> weights,
+    TConstArrayRef<float> sample,
+    TConstArrayRef<float> weights,
     const double alpha,
     const double delta
 ) {
